requirements.txt
streamlit>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
twilio>=8.5.0
python-dotenv
twilio>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dateutil>=2.8.2
jinja2>=3.1.2
aiofiles>=23.1.0
